Created simple.py
In order to familiarise with reinforcement learning with neural networks I created a much simplified version of a card game with only 2 players and 4 cards.

The players have 2 cards each, so the feature vector is 8 dimensional (4 for the hand and 4 for the table).

The script has some problem which probably has to do with something not related to the neural net. I upload it anyway because maybe some of you guys can give it a look and make it work, and then apply the ideas to the Klaverjassen script!

This should be the blueprint for the more complex Klaverjassen AI.