Added some result gathering
Now flow.py produces two plots, one showing the absolute reward (summed cumulatively) and the other the reward relative to the total value of the player's hand.  For details on this last implementation see table.DealCards and table.WhoWinsTrick.

I also added that, in addition to the team having a +1 reward for winning the trick, the player who actually wins the trick has an additional +0.2 reward.